IS-2024: Remove unused folder + refactor
