--- src/ikoner/LederIkon.js
@@ -1,29 +1,0 @@
-import React from "react";
-
-const LederIkon = () => {
-  return (
-    <svg
-      contentScriptType="text/ecmascript"
-      width="18px"
-      contentStyleType="text/css"
-      viewBox="0 0 18 18"
-      height="18px"
-      preserveAspectRatio="xMidYMid meet"
-      xmlns="http://www.w3.org/2000/svg"
-      version="1.0"
-    >
-      <g
-        fill="none"
-        strokeMiterlimit="10"
-        stroke="#3e3832"
-        strokeLinejoin="round"
-      >
-        <path d="M9.5 14.281v2.719l-5.009 1.789c-1.194.427-1.991 1.558-1.991 2.825v1.886h19v-1.886c0-1.268-.797-2.398-1.991-2.825l-5.009-1.789v-2.906" />
-        <ellipse rx="5" ry="6" cx="11.875" cy="9" />
-        <path d="M16.828 8.453l-.453.047c-1.703.328-2.797-.289-3.734-1.93-.563 1.078-2.322 1.93-3.766 1.93-.711 0-1.323-.146-1.936-.466" />
-      </g>
-    </svg>
-  );
-};
-
-export default LederIkon;
