--- src/ikoner/MoteIkon.js
@@ -1,36 +1,0 @@
-import React from "react";
-
-const MoteIkon = () => {
-  return (
-    <svg width="18px" height="18px" viewBox="0 0 18 18" version="1.1">
-      <g
-        id="HISTORIKK"
-        stroke="none"
-        strokeWidth="1"
-        fill="none"
-        fillRule="evenodd"
-      >
-        <g
-          id="Historikk_oppfølgingsplan-Copy"
-          transform="translate(-370.000000, -762.000000)"
-          stroke="#000000"
-        >
-          <g id="Møte_ikon" transform="translate(371.000000, 763.000000)">
-            <g id="Group">
-              <polyline
-                id="Shape"
-                points="3 1.66666667 0.333333333 1.66666667 0.333333333 15.6666667 15.6666667 15.6666667 15.6666667 1.66666667 13 1.66666667"
-              />
-              <path
-                d="M3,0.333333333 L5,0.333333333 L5,3 L3,3 L3,0.333333333 Z M11,0.333333333 L13,0.333333333 L13,3 L11,3 L11,0.333333333 Z M5,1.66666667 L11,1.66666667 L5,1.66666667 Z M0.333333333,5 L15.6666667,5 L0.333333333,5 Z"
-                id="Shape"
-              />
-            </g>
-          </g>
-        </g>
-      </g>
-    </svg>
-  );
-};
-
-export default MoteIkon;
