--- src/ikoner/OppfolgingsplanIkon.js
@@ -1,49 +1,0 @@
-import React from "react";
-
-const OppfolgingsplanIkon = () => {
-  return (
-    <svg width="18px" height="18px" viewBox="0 0 18 18" version="1.1">
-      <g
-        id="HISTORIKK"
-        stroke="none"
-        strokeWidth="1"
-        fill="none"
-        fillRule="evenodd"
-        strokeLinejoin="round"
-      >
-        <g
-          id="Historikk_oppfølgingsplan-Copy"
-          transform="translate(-371.000000, -838.000000)"
-          stroke="#000000"
-        >
-          <g
-            id="Oppfølgings_ikon"
-            transform="translate(372.000000, 839.000000)"
-          >
-            <g id="Group-5" strokeLinecap="round">
-              <polyline
-                id="Stroke-1"
-                points="9 9.66666667 5.66666667 9.66666667 3 12.3333333 3 9.66666667 0.333333333 9.66666667 0.333333333 0.333333333 14.3333333 0.333333333 14.3333333 6"
-              />
-              <polygon
-                id="Stroke-3"
-                points="10.6666667 15 8.33333333 15.6666667 9 13.3333333 14 8.33333333 15.6666667 10"
-              />
-            </g>
-            <path
-              d="M12.6666667,9.66666667 L14.3333333,11.3333333"
-              id="Stroke-6"
-            />
-            <path
-              d="M9,13.3333333 L10.6666667,15"
-              id="Stroke-7"
-              strokeLinecap="round"
-            />
-          </g>
-        </g>
-      </g>
-    </svg>
-  );
-};
-
-export default OppfolgingsplanIkon;
