--- src/navigation/menypunkterTypes.ts
@@ -1,17 +1,0 @@
-export enum Menypunkter {
-  AKTIVITETSKRAV = "AKTIVITETSKRAV",
-  BEHANDLERDIALOG = "BEHANDLERDIALOG",
-  DIALOGMOTE = "DIALOGMOTE",
-  NOKKELINFORMASJON = "NOKKELINFORMASJON",
-  SYKMELDINGER = "SYKMELDINGER",
-  SYKEPENGESOKNADER = "SYKEPENGESOKNADER",
-  OPPFOELGINGSPLANER = "OPPFOELGINGSPLANER",
-  HISTORIKK = "HISTORIKK",
-  VEDTAK = "VEDTAK",
-}
-
-export type Menypunkt = {
-  navn: string;
-  sti: string;
-  menypunkt: Menypunkter;
-};
