+++ src/sider/behandlerdialog/BehandlerdialogContainer.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement } from "react";
 import Side from "@/sider/Side";
-import { Menypunkter } from "@/navigation/menypunkterTypes";
 import Sidetopp from "@/components/Sidetopp";
 import SideLaster from "@/components/SideLaster";
 import { BehandlerdialogSide } from "@/sider/behandlerdialog/BehandlerdialogSide";
+import { Menypunkter } from "@/components/globalnavigasjon/GlobalNavigasjon";
 
 const texts = {
   title: "Dialog med behandler",
