+++ src/sider/sykmeldinger/container/SykmeldingerContainer.tsx
@@ -4,15 +4,9 @@ import DineSykmeldinger from "../sykmeldinger/DineSykmeldinger";
 import Pengestopp from "../../../components/pengestopp/Pengestopp";
 import SideLaster from "../../../components/SideLaster";
 import { useSykmeldingerQuery } from "@/data/sykmelding/sykmeldingQueryHooks";
-import { Menypunkter } from "@/navigation/menypunkterTypes";
 import { BistandsbehovOppgaver } from "@/sider/sykmeldinger/VurderBistandsbehov";
 import { Heading } from "@navikt/ds-react";
-import styled from "styled-components";
-
-const StyledHeading = styled(Heading)`
-  text-align: center;
-  margin: 1em 0 0.5em 0;
-`;
+import { Menypunkter } from "@/components/globalnavigasjon/GlobalNavigasjon";
 
 const SykmeldingerSide = (): ReactElement => {
   const { isLoading, isError, sykmeldinger } = useSykmeldingerQuery();
@@ -23,7 +17,9 @@ const SykmeldingerSide = (): ReactElement => {
         <div>
           <Pengestopp sykmeldinger={sykmeldinger} />
           <BistandsbehovOppgaver />
-          <StyledHeading size="xlarge">Sykmeldinger</StyledHeading>
+          <Heading size="xlarge" className="text-center mt-4 mb-2">
+            Sykmeldinger
+          </Heading>
           <DineSykmeldinger />
         </div>
       </SideLaster>
